Refactor xml
