--- taxel-xml/src/write_tags.rs
@@ -1,3 +1,0 @@
-use crate::{CsvRow, Tag};
-use csv::Writer;
-use log::info;
